pyproject.toml
[project]
name = "langchain-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-anthropic>=0.3.21",
    "langchain-experimental>=0.3.4",
    "langchain-tavily>=0.2.11",
    "langchain[anthropic,openai]>=0.3.27",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.6.8",
    "langgraph-supervisor>=0.0.29",
    "matplotlib>=3.10.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pymongo>=4.6.0",
    "motor>=3.6.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.38.0",
    "pymilvus>=2.4.0",
    "chromadb>=0.4.0",
    "pydantic>=2.0.0",
    "aiofiles>=25.1.0",
    "python-dotenv>=1.2.1",
    "python-louvain>=0.16",
    "networkx>=3.5",
    "pyarrow>=22.0.0",
    "graphrag>=0.1.0",
    "scalar-fastapi>=1.0.6",
    "langchain-google-vertexai>=0.0.1",
    "tavily-python>=0.7.14",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
